+++ R/cGroupsGUI.R
@@ -943,7 +943,8 @@ cGroupsGUI <- function(X){
                                               max.xlev=10, timemax=NA, Q1=0.25, Q3=0.75, simplify=TRUE, ref.no=NA, fact.ratio=1, p.corrected=TRUE,
                                               compute.ratio=TRUE, include.miss=FALSE, oddsratio.method='midp',chisq.test.perm = FALSE, 
                                               byrow = FALSE, chisq.test.B = 2000, chisq.test.seed = NULL, 
-                                              Date.format = 'd-mon-Y', var.equal = TRUE)"))))
+                                              Date.format = 'd-mon-Y', var.equal = TRUE, 
+                                              conf.level=0.95, surv=FALSE, riskratio=FALSE, riskratio.method='wald')"))))
             if(inherits(ans,"try-error")) return(tcltk::tkmessageBox(message = "Error occurred in 'compareGroups' function", icon = "info", type="ok"))
             print(ans)
             if(!inherits(ans,"try-error")) ans <- try(createTable(x=ans,hide=res$hide, digits=res$digits,type=res$type, show.all= res$show.all, show.p.trend= res$show.p.trend,show.p.mul= res$show.p.mul,show.n=res$show.n, show.p.overall = res$show.p.overall, show.ratio = res$show.ratio, show.descr = res$show.descr))
@@ -959,7 +960,8 @@ cGroupsGUI <- function(X){
                                               max.xlev=10, timemax=NA, Q1=0.25, Q3=0.75, simplify=TRUE, ref.no=NA, fact.ratio=1, p.corrected=TRUE,
                                               compute.ratio=TRUE, include.miss=FALSE, oddsratio.method='midp',chisq.test.perm = FALSE, 
                                               byrow = FALSE, chisq.test.B = 2000, chisq.test.seed = NULL, 
-                                              Date.format = 'd-mon-Y', var.equal = TRUE)"))))
+                                              Date.format = 'd-mon-Y', var.equal = TRUE, 
+                                              conf.level=0.95, surv=FALSE, riskratio=FALSE, riskratio.method='wald')"))))
             if(inherits(ans,"try-error")) return(tcltk::tkmessageBox(message = "Error occurred in 'compareGroups' function", icon = "info", type="ok"))
             if(!inherits(ans,"try-error")) ans <- try(createTable(x=ans,hide=res$hide, digits=res$digits,type=res$type, show.all= res$show.all, show.p.trend= res$show.p.trend,show.p.mul= res$show.p.mul,show.n=res$show.n, show.p.overall = res$show.p.overall, show.ratio = res$show.ratio, show.descr = res$show.descr))
             if(inherits(ans,"try-error")) return(tcltk::tkmessageBox(message = "Error occurred in 'createTable' function", icon = "info", type="ok"))
@@ -985,7 +987,8 @@ cGroupsGUI <- function(X){
                                           max.xlev=10, timemax=NA, Q1=0.25, Q3=0.75, simplify=TRUE, ref.no=NA, fact.ratio=1, p.corrected=TRUE,
                                           compute.ratio=TRUE, include.miss=FALSE, oddsratio.method='midp',chisq.test.perm = FALSE, 
                                           byrow = FALSE, chisq.test.B = 2000, chisq.test.seed = NULL, 
-                                          Date.format = 'd-mon-Y', var.equal = TRUE)"))))
+                                          Date.format = 'd-mon-Y', var.equal = TRUE,
+                                          conf.level=0.95, surv=FALSE, riskratio=FALSE, riskratio.method='wald')"))))
         if(inherits(ans,"try-error")) return(tcltk::tkmessageBox(message = "Error occurred in 'compareGroups' function", icon = "info", type="ok"))
             if(!inherits(ans,"try-error")) ans <- try(createTable(x=ans,hide=res$hide, digits=res$digits,type=res$type, show.all= res$show.all, show.p.trend= res$show.p.trend,show.p.mul= res$show.p.mul,show.n=res$show.n, show.p.overall = res$show.p.overall, show.ratio = res$show.ratio, show.descr = res$show.descr))
         if(inherits(ans,"try-error")) return(tcltk::tkmessageBox(message = "Error occurred in 'createTable' function", icon = "info", type="ok"))
@@ -1011,7 +1014,8 @@ cGroupsGUI <- function(X){
                                              max.xlev=10, timemax=NA, Q1=0.25, Q3=0.75, simplify=TRUE, ref.no=NA, fact.ratio=1, p.corrected=TRUE,
                                              compute.ratio=TRUE, include.miss=FALSE, oddsratio.method='midp',chisq.test.perm = FALSE, 
                                              byrow = FALSE, chisq.test.B = 2000, chisq.test.seed = NULL, 
-                                             Date.format = 'd-mon-Y', var.equal = TRUE)"))))
+                                             Date.format = 'd-mon-Y', var.equal = TRUE, 
+                                             conf.level=0.95, surv=FALSE, riskratio=FALSE, riskratio.method='wald')"))))
           if(inherits(ans,"try-error")) return(tcltk::tkmessageBox(message = "Error occurred in 'compareGroups' function", icon = "info", type="ok"))
             if(!inherits(ans,"try-error")) ans <- try(createTable(x=ans,hide=res$hide, digits=res$digits,type=res$type, show.all= res$show.all, show.p.trend= res$show.p.trend,show.p.mul= res$show.p.mul,show.n=res$show.n, show.p.overall = res$show.p.overall, show.ratio = res$show.ratio, show.descr = res$show.descr))
           if(inherits(ans,"try-error")) return(tcltk::tkmessageBox(message = "Error occurred in 'createTable' function", icon = "info", type="ok"))
@@ -1037,7 +1041,8 @@ cGroupsGUI <- function(X){
                                               max.xlev=10, timemax=NA, Q1=0.25, Q3=0.75, simplify=TRUE, ref.no=NA, fact.ratio=1, p.corrected=TRUE,
                                               compute.ratio=TRUE, include.miss=FALSE, oddsratio.method='midp',chisq.test.perm = FALSE, 
                                               byrow = FALSE, chisq.test.B = 2000, chisq.test.seed = NULL, 
-                                              Date.format = 'd-mon-Y', var.equal = TRUE)"))))
+                                              Date.format = 'd-mon-Y', var.equal = TRUE, 
+                                              conf.level=0.95, surv=FALSE, riskratio=FALSE, riskratio.method='wald')"))))
           if(inherits(ans,"try-error")) return(tcltk::tkmessageBox(message = "Error occurred in 'compareGroups' function", icon = "info", type="ok"))
             if(!inherits(ans,"try-error")) ans <- try(createTable(x=ans,hide=res$hide, digits=res$digits,type=res$type, show.all= res$show.all, show.p.trend= res$show.p.trend,show.p.mul= res$show.p.mul,show.n=res$show.n, show.p.overall = res$show.p.overall, show.ratio = res$show.ratio, show.descr = res$show.descr))
           if(inherits(ans,"try-error")) return(tcltk::tkmessageBox(message = "Error occurred in 'createTable' function", icon = "info", type="ok"))