+++ R/export2md.R
@@ -89,7 +89,7 @@ export2md<-function(x, which.table="descr", nmax=TRUE, header.labels=c(), captio
       table1[((1+n.exists):nrow(table1)),ncol(table1)] <- ifelse(table1[((1+n.exists):nrow(table1)),ncol(table1)]=="", "\\vphantom{}", table1[((1+n.exists):nrow(table1)),ncol(table1)])
     if (format=="latex") caption <- gsub("%","\\\\%",caption)
     ans <- knitr::kable(table1, align = align, row.names = FALSE, caption=caption[1], format=format, 
-                        booktabs=format=="latex", longtable=FALSE, linesep="", ...)
+                        booktabs=format=="latex", longtable=TRUE, linesep="", ...)
     ans <- add_indent(ans, grep("^ ",table1[,1]))
     if (width!=Inf) ans <- column_spec(ans, 1, width = width)
     # groups    
@@ -142,7 +142,7 @@ export2md<-function(x, which.table="descr", nmax=TRUE, header.labels=c(), captio
     align <- c("l", rep("c", ncol(table2)))
     colnames(table2)[-1] <- trim(table2[1, -1])
     table2 <- table2[-1, ,drop=FALSE]
-    ans <- knitr::kable(table2, align = align, row.names = FALSE, caption=caption[1], format=format, booktabs=format=="latex", longtable=FALSE, ...)
+    ans <- knitr::kable(table2, align = align, row.names = FALSE, caption=caption[1], format=format, booktabs=format=="latex", longtable=TRUE, ...)
     # ans <- knitr::kable(table2, align = align, row.names = FALSE, caption=caption[1], format=format, booktabs=format=="latex")
     # groups    
     if (!is.null(cc)){